% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hajek_est_function.R
\name{hajek_est}
\alias{hajek_est}
\title{Hajék estimator with sandwich-type SEs}
\usage{
hajek_est(data, w, y, z, block)
}
\arguments{
\item{data}{a dataframe of weights, outcomes, treatment assignments, and block indicators}

\item{y}{a character, the name of the outcome column}

\item{z}{a character, the name of the treatment indicator column}

\item{block}{a character, the name of the block indicator column}

\item{weight}{a character, the name of the weight column}
}
\value{
a list of results. $theta: a vector of Hajék ratios; $varest: a matrix of varaince estimates, $covest: an array of covariance bound estimates.
}
\description{
This function takes a dataframe and returns the Hajék estimator with
3 or 1 variance estimator and 4 or 3 covariance estimator depending on the
block and treatment group sizes
}
