% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ss_function.R
\name{mean_ss}
\alias{mean_ss}
\title{A function for computing the variance estimation}
\usage{
mean_ss(a, b)
}
\arguments{
\item{a}{a vector of data}

\item{b}{a vector of data}
}
\value{
The value of mean((a[i]-b[j])^2) taken over (i,j)'s.
}
\description{
This function takes two vectors and returns an average of squared differences.
}
