% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scd_term_function.R
\name{scd_term}
\alias{scd_term}
\title{A function for covariance estimation (2nd term)}
\usage{
scd_term(a, nb)
}
\arguments{
\item{a}{a vector of data, gamma's of a block}

\item{nb}{an integer, the block size}
}
\value{
the value of nb * (mean(a[i]^2) + mean(a[i]a[j])/(nb-1))
}
\description{
This function takes a vector and an integer and returns the second
term of the covariance bound estimator
}
