% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thd_term_1_function.R
\name{thd_term_1}
\alias{thd_term_1}
\title{A function for covariance estimation (one component of the 3rd term)}
\usage{
thd_term_1(a, b, pil, pim, nb)
}
\arguments{
\item{a}{a vector of data, gamma' of a treatment group l}

\item{b}{a vector of data, gamma' of another treatment group m}

\item{pil}{a real number, the propensity score of treatment l}

\item{pim}{a real number, the propensity score of treatment m}

\item{nb}{an integer, the block size}
}
\value{
the value of mean(a[i]b[j]) / pil / pim
}
\description{
This function takes a few arguments and compute cone compunent of the
third term of the covariance bound estimator
}
